adjust structure of the file
